requirements.txt
aiogram
aiohttp
aioredis
python-dotenv
loguru
peewee
aiohttp_socks
loguru
